move HTTPUrl and Color out to be generic pydantic types
