+++ src/pyticktick/models/__init__.py
@@ -1,0 +1,3 @@
+from pyticktick.models.pydantic import Color, HttpUrl
+
+__all__ = ["Color", "HttpUrl"]
