+++ src/pyticktick/models/v2/parameters/project.py
@@ -11,7 +11,8 @@ from typing import Literal, Union
 
 from pydantic import BaseModel, ConfigDict, Field
 
-from pyticktick.models.v2.types import Color, ObjectId
+from pyticktick.models.pydantic import Color
+from pyticktick.models.v2.types import ObjectId
 
 
 class CreateProjectV2(BaseModel):
